refactor: consolidate applicant model methods and improve blacklisted applicant retrieval
